AMP-18483: IntWrapper inherited a field 'value' and had one declared of its own -- remove the declared one
git-svn-id: file:///svn/branches/AMP_2_10_RELEASE@25717 9e8d75aa-d0cb-44a9-8341-a4b8ba1a38b4

+++ amp/WEB-INF/src/org/dgfoundation/amp/ar/view/xls/XLSExporter.java
@@ -126,7 +126,7 @@ public abstract class XLSExporter extends Exporter {
 	}
 
 	protected HSSFCell getCell(HSSFRow row, HSSFCellStyle style) {
-		HSSFCell cell = row.createCell((short)colId.intValue());
+		HSSFCell cell = row.createCell((int)colId.intValue());
 //		HSSFCellStyle cellstyle = wb.createCellStyle();
 //		cellstyle.cloneStyleFrom(style);
 		cell.setCellStyle(style);
