+++ amp/WEB-INF/src/org/digijava/module/aim/action/XLSExportAction.java
@@ -234,7 +234,7 @@ public class XLSExportAction extends Action {
 		rowId.inc();
 		colId.reset();
 		row=sheet.createRow(rowId.shortValue());
-		HSSFCell cell=row.createCell(colId.shortValue());
+		HSSFCell cell=row.createCell(colId.intValue());
 		if(reportForm!=null && reportForm.getLogoOptions() !=null)
 			if (reportForm.getLogoOptions().equals("0")) {//disabled
 				// do nothing
